Crate cloudmonitoring2_beta2

This documentation was generated from Cloud Monitoring crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

metric descriptors (create, delete and list)

timeseries (list and write)

timeseries descriptors (list)

Everything else about the Cloud Monitoring v2_beta2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.metric_descriptors().list(...).doit()
let r = hub.metric_descriptors().delete(...).doit()
let r = hub.metric_descriptors().create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
cloudmonitoring2_beta2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate cloudmonitoring2_beta2;
use cloudmonitoring2_beta2::ListMetricDescriptorsRequest;
use cloudmonitoring2_beta2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use cloudmonitoring2_beta2::CloudMonitoring;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudMonitoring::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ListMetricDescriptorsRequest = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.metric_descriptors().list(&req, "project")
             .query("dolores")
             .page_token("kasd")
             .count(-22)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

CloudMonitoring
Central instance to access all CloudMonitoring related resource activities

DeleteMetricDescriptorResponse
The response of cloudmonitoring.metricDescriptors.delete.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete metric descriptors (response)

ListMetricDescriptorsRequest
The request of cloudmonitoring.metricDescriptors.list.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list metric descriptors (request)

ListMetricDescriptorsResponse
The response of cloudmonitoring.metricDescriptors.list.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list metric descriptors (response)

ListTimeseriesDescriptorsRequest
The request of cloudmonitoring.timeseriesDescriptors.list

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries descriptors (request)

ListTimeseriesDescriptorsResponse
The response of cloudmonitoring.timeseriesDescriptors.list

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries descriptors (response)

ListTimeseriesRequest
The request of cloudmonitoring.timeseries.list

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries (request)

ListTimeseriesResponse
The response of cloudmonitoring.timeseries.list

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries (response)

MetricDescriptor
A metricDescriptor defines the name, label keys, and data type of a particular metric.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list metric descriptors (none)

delete metric descriptors (none)

create metric descriptors (request|response)

MetricDescriptorCreateMethodBuilder
Create a new metric.

MetricDescriptorDeleteMethodBuilder
Delete an existing metric.

MetricDescriptorLabelDescriptor
A label in a metric is a description of this metric, including the key of this description (what the description is), and the value for this description.

This type is not used in any activity, and only used as part of another schema.

MetricDescriptorListMethodBuilder
List metric descriptors that match the query. If the query is not set, then all of the metric descriptors will be returned. Large responses will be paginated, use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

MetricDescriptorMethodsBuilder
A builder providing access to all methods supported on metricDescriptor resources. It is not used directly, but through the CloudMonitoring hub.

MetricDescriptorTypeDescriptor
A type in a metric contains information about how the metric is collected and what its data points look like.

This type is not used in any activity, and only used as part of another schema.

Point
Point is a single point in a time series. It consists of a start time, an end time, and a value.

This type is not used in any activity, and only used as part of another schema.

PointDistribution
Distribution data point value type. When writing distribution points, try to be consistent with the boundaries of your buckets. If you must modify the bucket boundaries, then do so by merging, partitioning, or appending rather than skewing them.

This type is not used in any activity, and only used as part of another schema.

PointDistributionBucket
The histogram's bucket. Buckets that form the histogram of a distribution value. If the upper bound of a bucket, say U1, does not equal the lower bound of the next bucket, say L2, this means that there is no event in [U1, L2).

This type is not used in any activity, and only used as part of another schema.

PointDistributionOverflowBucket
The overflow bucket is a special bucket that does not have the upperBound field; it includes all of the events that are no less than its lower bound.

This type is not used in any activity, and only used as part of another schema.

PointDistributionUnderflowBucket
The underflow bucket is a special bucket that does not have the lowerBound field; it includes all of the events that are less than its upper bound.

This type is not used in any activity, and only used as part of another schema.

Timeseries
The monitoring data is organized as metrics and stored as data points that are recorded over time. Each data point represents information like the CPU utilization of your virtual machine. A historical record of these data points is called a time series.

This type is not used in any activity, and only used as part of another schema.

TimeseriesDescriptor
TimeseriesDescriptor identifies a single time series.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries descriptors (none)

TimeseriesDescriptorLabel
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

TimeseriesDescriptorListMethodBuilder
List the descriptors of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

TimeseriesDescriptorMethodsBuilder
A builder providing access to all methods supported on timeseriesDescriptor resources. It is not used directly, but through the CloudMonitoring hub.

TimeseriesPoint
When writing time series, TimeseriesPoint should be used instead of Timeseries, to enforce single point for each time series in the timeseries.write request.

This type is not used in any activity, and only used as part of another schema.

TimeseryListMethodBuilder
List the data points of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

TimeseryMethodsBuilder
A builder providing access to all methods supported on timesery resources. It is not used directly, but through the CloudMonitoring hub.

TimeseryWriteMethodBuilder
Put data points to one or more time series for one or more metrics. If a time series does not exist, a new time series will be created. It is not allowed to write a time series point that is older than the existing youngest point of that time series. Points that are older than the existing youngest point of that time series will be discarded silently. Therefore, users should make sure that points of a time series are written sequentially in the order of their end time.

WriteTimeseriesRequest
The request of cloudmonitoring.timeseries.write

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

write timeseries (request)

WriteTimeseriesResponse
The response of cloudmonitoring.timeseries.write

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

write timeseries (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.